remove OpticCan constraint whenever possible
