+++ src/main/scala/monocly/OpticCan.scala
@@ -1,0 +1,22 @@
+package monocly
+
+sealed trait OpticCan
+
+//        GetMany
+//          /\
+//         /  \
+// GetOption  GetOneOrMore
+//         \  /
+//          \/
+//        GetOne
+trait GetMany extends OpticCan
+trait GetOption extends GetMany
+trait GetOneOrMore extends GetMany
+trait GetOne extends GetOption with GetOneOrMore
+
+//    Modify
+//      ^
+//      |
+//  ReverseGet
+trait Modify extends OpticCan
+trait ReverseGet extends Modify
