+++ src/main/scala/monocly/impl/GetterImpl.scala
@@ -4,9 +4,9 @@ import monocly.internal._
 import monocly._
 
 
-trait GetterImpl[+ThisCan <: OpticCan, -S, +A]:
+trait GetterImpl[+ThisCan, -S, +A]:
 
-  def canAlso[NewCan <: OpticCan]: GetterImpl[NewCan, S, A] = 
+  def canAlso[NewCan <: OpticCan]: GetterImpl[NewCan, S, A] =
     asInstanceOf[GetterImpl[NewCan, S, A]]
 
   def preComposeGetMany[ThatCan <: GetMany, S0](impl1: GetManyImpl[ThatCan, S0, S]): GetterImpl[ThisCan | ThatCan, S0, A]
@@ -14,7 +14,7 @@ trait GetterImpl[+ThisCan <: OpticCan, -S, +A]:
   def preComposeGetOption[ThatCan <: GetOption, S0](impl1: GetOptionImpl[ThatCan, S0, S]): GetterImpl[ThisCan | ThatCan, S0, A]
   def preComposeGetOne[ThatCan <: GetOne, S0](impl1: GetOneImpl[ThatCan, S0, S]): GetterImpl[ThisCan | ThatCan, S0, A]
 
-  def andThen[ThatCan <: OpticCan, C](impl2: GetterImpl[ThatCan, A, C]): GetterImpl[ThisCan | ThatCan, S, C]
+  def andThen[ThatCan, C](impl2: GetterImpl[ThatCan, A, C]): GetterImpl[ThisCan | ThatCan, S, C]
 
   def get(using ThisCan <:< GetOne): S => A = sys.error("This optic does not support 'get'")
   def getOption(using ThisCan <:< GetOption): S => Option[A] = sys.error("This optic does not support 'getOption'")
