+++ src/main/scala/monocly/impl/ModifyImpl.scala
@@ -10,7 +10,7 @@ class ModifyImpl[ThisCan <: Modify, -S, +T, +A, -B](_modify: (A => B) => S => T)
   override def preComposeReverseGet[ThatCan <: ReverseGet, S0, T0](impl1: ReverseGetImpl[ThatCan, S0, T0, S, T]): ModifyImpl[ThisCan | ThatCan, S0, T0, A, B] = 
     ModifyImpl(f => s0 => impl1.modify(s => modify(f)(s))(s0))
 
-  override def andThen[ThatCan <: OpticCan, C, D](impl2: SetterImpl[ThatCan, A, B, C, D]): SetterImpl[ThisCan | ThatCan, S, T, C, D] = 
+  override def andThen[ThatCan, C, D](impl2: SetterImpl[ThatCan, A, B, C, D]): SetterImpl[ThisCan | ThatCan, S, T, C, D] =
     impl2.preComposeModify(this)
 
   override def modify(f: A => B)(using ThisCan <:< Modify): S => T = _modify(f)
