+++ src/main/scala/monocly/impl/NoGetter.scala
@@ -7,5 +7,5 @@ object NoGetter extends GetterImpl[Nothing, Any, Nothing]:
   override def preComposeGetOneOrMore[ThatCan <: GetOneOrMore, S0](impl1: GetOneOrMoreImpl[ThatCan, S0, Any]) = NoGetter
   override def preComposeGetOption[ThatCan <: GetOption, S0](impl1: GetOptionImpl[ThatCan, S0, Any]) = NoGetter
   override def preComposeGetOne[ThatCan <: GetOne, S0](impl1: GetOneImpl[ThatCan, S0, Any]) = NoGetter
-  override def andThen[ThatCan <: OpticCan, C](impl2: GetterImpl[ThatCan, Nothing, C]) = NoGetter
+  override def andThen[ThatCan, C](impl2: GetterImpl[ThatCan, Nothing, C]) = NoGetter
 end NoGetter
